// src/master/product/interface.ts
export interface ISalesman {
  emp_id: number;
  emp_code: string;
  emp_name: string;
  sales_team_id: number;
  sales_team_code: string;
  sales_team_name: string;
  sales_name: string;
  opr_type: string;
  opr_type_name: string;
  is_bonus_rep: boolean;
  trans_date: null;
  wh_id: number;
  wh_code: string;
  wh_name: string;
  inc_grp_id: number;
  official_id: number;
  official_type: number;
  hierarchy_code: string;
  sale_system: string;
  sm_is_transfer: boolean;
  sm_valid_route: boolean;
  sm_geoloc_valid: boolean;
  sm_radius: number;
  sm_password: string;
  is_active: boolean;
  updated_at: Date;
  updated_by_name: string;
  image_url: string;
  sm_is_barcode: boolean;
  sm_is_photo_profile: boolean;
}

export interface IOutlet {
  outlet_id: number;
  outlet_code: string;
  outlet_name: string;
  barcode: string;
  outlet_status: number;
  address1: string;
  address2: string;
  city: string;
  zip_code: string;
  phone_no: string;
  wa_no: string;
  fax_no: string;
  email: string;
  disc_grp_id: number;
  disc_grp_code: string;
  disc_grp_name: string;
  ot_loc_id: number;
  ot_loc_code: string;
  ot_loc_name: string;
  ot_grp_id: number;
  ot_grp_code: string;
  ot_grp_name: string;
  price_grp_id: number;
  price_grp_code: string;
  price_grp_name: string;
  district_id: number;
  district_code: string;
  district_name: string;
  beat_id: number;
  beat_code: string;
  beat_name: string;
  sbeat_id: number;
  sbeat_code: string;
  sbeat_name: string;
  ot_class_id: number;
  ot_class_code: string;
  ot_class_name: string;
  industry_id: number;
  industry_code: string;
  industry_name: string;
  market_id: number;
  market_code: string;
  market_name: string;
  top: number;
  payment_type: number;
  is_contra_bon: boolean;
  plu_grp_id: number;
  plu_grp_code: string;
  plu_grp_name: string;
  conv_grp_id: number;
  conv_grp_code: string;
  conv_grp_name: string;
  disc_inv_id: number;
  disc_inv_code: string;
  disc_inv_name: string;
  agent_from: string;
  credit_limit_type: number;
  credit_limit: number;
  sales_inv_limit_type: number;
  sales_inv_limit: number;
  avg_sales_week: number;
  avg_sales_month: number;
  first_trans_date: string;
  last_trans_date: string;
  first_week_no: number;
  ot_start_date: string;
  ot_reg_date: string;
  building_own: number;
  dob: string;
  ar_status: number;
  ar_total: number;
  closed_date: string;
  is_emb_bail: boolean;
  tax_name: string;
  tax_addr1: string;
  tax_addr2: string;
  tax_city: string;
  tax_no: string;
  tax_invoice_form: string;
  owner_name: string;
  owner_addr1: string;
  owner_addr2: string;
  owner_city: string;
  owner_phone_no: string;
  owner_id_no: string;
  delv_addr1: string;
  delv_addr2: string;
  delv_city: string;
  inv_addr1: string;
  inv_addr2: string;
  inv_city: string;
  is_active: boolean;
  updated_by: number;
  updated_at: Date;
  updated_by_name: string;
  latitude: string;
  longitude: string;
  ot_type_id: number;
  ot_type_code: null;
  ot_type_name: null;
  is_obs: boolean;
  obs: string;
  outlet_ward_id: string;
  outlet_ward: string;
  outlet_sub_district_id: string;
  outlet_sub_district: string;
  outlet_regency_id: string;
  outlet_regency: string;
  outlet_province_id: string;
  outlet_province: string;
  is_wa_no: null;
  delv_ward_id: null;
  delv_ward: null;
  delv_sub_district_id: null;
  delv_sub_district: null;
  delv_regency_id: null;
  delv_regency: null;
  delv_province_id: null;
  delv_province: null;
  delv_zip_code: null;
  delv_is_same_addr: null;
  inv_ward_id: null;
  inv_ward: null;
  inv_sub_district_id: null;
  inv_sub_district: null;
  inv_regency_id: null;
  inv_regency: null;
  inv_province_id: null;
  inv_province: null;
  inv_zip_code: null;
  inv_is_same_addr: null;
  verification_status: null;
}

export interface IQuantityProduct {
  name: string
  qty: number
  price_ppn: number
  price: number
}